% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo_read_cypher.R
\name{neo_read_cypher}
\alias{neo_read_cypher}
\title{Read in a cypher file}
\usage{
neo_read_cypher(cypher_file)
}
\arguments{
\item{cypher_file}{character. The path to the cypher file to be read into memory.}
}
\description{
Read a cypher file that exists on disk into a character vector of length 1.
}
