% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo_start.R
\name{neo_start}
\alias{neo_start}
\title{If Neo4j is not running, start the database}
\usage{
neo_start(neo_path = "~/neo4j-community-3.2.0")
}
\arguments{
\item{neo_path}{character. The Quorum HOST URL}
}
\description{
This function will check to see if Neo4j is running, and if not, it
will kick off the database.  ASSUMES that you have command line access.
}
