% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo_wipe.R
\name{neo_wipe}
\alias{neo_wipe}
\title{Totally wipe out the default Neo4j database data.}
\usage{
neo_wipe(neo_path = "~/neo4j-community-3.2.0")
}
\arguments{
\item{neo_path}{character. The path for the neo4j install}

\item{cql}{character. A path to the cypher file which can contain multiple commands to}
}
\description{
Function to stop a Neo4j server, if running, and totally rebuild the default database from scratch. This is
different from wiping the database via commands but keeping the server running.
Why?:  Replicate a fresh database instance, which is helpful for developing and certain analytic instances.
}
