% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo_run_cypher.R
\name{neo_run_cypher}
\alias{neo_run_cypher}
\title{Using the shell, import a cypher file you have on disk}
\usage{
neo_run_cypher(neo_path = "~/neo4j-community-3.2.0", cql)
}
\arguments{
\item{neo_path}{character. The path for the neo4j install}

\item{cql}{character. A path to the cypher file which can contain multiple commands to}
}
\description{
Import data into a running neo4j server using cypher via the shell import tool
ASSUMES that you have command line access.
}
